Setting up modular arithmetic for manip of chars.
Need to be able to send any alphabetic character,
as well as a null (used @ in this version, due to its
position in ascii, but can be hidden behind a simpler
abstraction).  Need to be able to handle multiple
keys/keyops for a single encoding, as the number of
keys shared between clients is variable.