Setting the server to take a limited number of clients.
This is needed to ensure that the process of distributing
keys can actually occur.  A more useful way of doing
this would be if the client program/protocol had a
message to indicate that the last client is connecting.
The server could just wait for this message, then
start doing the key generation, assignment, 
rounds/message passing etc, since it knows that no
more clients are going to enter the picture.