Setting up the server-side FSM.
This currently adds some extra logic to the system
as a whole, and has resulting in more merging of
encryption code from the master branch to the current
active branch.

Also need to create a full cycle of client-server
communication to test the continued operation of the
cycle, so that it can be connected to the GUI and
have actual messages sent through by checking during
the FSM loops (at the start).