Added collision resolution.
This picks a random decay of min size 0, and max
size the length of the message that was being sent
(in order to ensure that giant delays don't happen.
This doesn't solve the problem that rounds are still
performed on characters only, so random delays mean
that messages are interleaved.  This is a problem
that does not really get mentioned in the text, so
it's probably a good idea to at least mention how
it comes to affect teh system in implementation,
even if it isn't fixed in this case.